refactor(chats): extract systemState + fuzzy search + textEdit tracking from useAiChat (Wave 2)
Move pure helpers out of the monolithic useAiChat hook into their own
modules. No behavior change — every helper keeps its existing signature
and the new modules are imported in place of the inline definitions.

New files
- src/apps/chats/utils/fuzzySearch.ts
    normalizeSearchText, computeMatchScore, deriveScoreThreshold
    (with module-private Levenshtein primitives).
- src/apps/chats/utils/systemState.ts
    getSystemState, getAssistantVisibleText, isChatsInForeground,
    showBackgroundedMessageNotification.
- src/apps/chats/utils/textEditInstanceTracking.ts
    recentlyCreatedTextEditInstances Map + trackNewTextEditInstance +
    getRecentTextEditInstanceForPath (write -> open race fallback).

Removed duplication
- karaokeHandler.ts had its own copy of normalizeSearchText / Levenshtein /
  computeMatchScore / deriveScoreThreshold with a 'should eventually be
  moved to helpers' comment. Now imports from utils/fuzzySearch.
- useRyoChat.ts had a second (legacy-shape) getSystemState — the
  /api/ai/ryo-reply endpoint only reads systemState.chatRoomContext
  (api/ai/ryo-reply.ts L22-27), so we can safely send the canonical
  payload. Drops ~90 lines and 5 store imports from the hook.
- detectUserOS is now sourced from tools/helpers.ts (the canonical
  copy used by handlers); useAiChat's local duplicate is gone.

useAiChat.ts shrinks from 2,745 -> 2,227 lines.

Tests
- tests/test-chats-fuzzy-search.test.ts (13 cases) — locks the matching
  algorithm so any future tweak is intentional.
- tests/test-chats-text-for-speech.test.ts (11 cases) — locks
  cleanTextForSpeech behavior (code fences, HTML, markdown links,
  bare URLs/email, whitespace collapsing, leading punctuation trim).

Refs plans/chats_useaichat_tts_cleanup.md Wave 2.
